+++ libcloud/test/compute/test_gce.py
@@ -27,7 +27,8 @@ from libcloud.compute.drivers.gce import (GCENodeDriver, API_VERSION,
                                           GCEAddress, GCEHealthCheck,
                                           GCEFirewall, GCEForwardingRule,
                                           GCENetwork,
-                                          GCEZone)
+                                          GCEZone,
+                                          GCENodeImage)
 from libcloud.common.google import (GoogleBaseAuthConnection,
                                     GoogleInstalledAppAuthConnection,
                                     GoogleBaseConnection,
@@ -241,6 +242,13 @@ class GCENodeDriverTest(LibcloudTestCase, TestCaseMixin):
         self.assertEqual(hc.interval, 10)
         self.assertEqual(hc.extra['host'], 'lchost')
         self.assertEqual(hc.extra['description'], 'test healthcheck')
+
+    def test_ex_create_image(self):
+        volume = self.driver.ex_get_volume('lcdisk')
+        image = self.driver.ex_create_image('coreos', volume)
+        self.assertTrue(isinstance(image, GCENodeImage))
+        self.assertEquals(image.name, 'coreos')
+        self.assertEquals(image.extra['description'], 'CoreOS test image')
 
     def test_ex_create_firewall(self):
         firewall_name = 'lcfirewall'